Add files via upload
I have looked through your code and did some manual unit testing. without edit, the numbers it returns for the lone_sum are not quite correct. one of the operands >= should have been a ==, the last elif statement should have been the if statement. I swapped those two statements around and now it works as expected. unit test with 1,1,1 =0 and 1,1,6=6 and 1,3,6=10 returned the correct results.
please review and let me know if I can be of further assistance.